Format and get build apk working

+++ packages/camera/camera_android_camerax/android/src/main/java/io/flutter/plugins/camerax/ZoomStateFlutterApiImpl.java
@@ -7,24 +7,30 @@ package io.flutter.plugins.camerax;
 import androidx.annotation.NonNull;
 import androidx.camera.core.ZoomState;
 import io.flutter.plugin.common.BinaryMessenger;
-import io.flutter.plugins.camerax.GeneratedCameraXLibrary.ZoomStateFlutterApi;;
+import io.flutter.plugins.camerax.GeneratedCameraXLibrary.ZoomStateFlutterApi;
 
 public class ZoomStateFlutterApiImpl extends ZoomStateFlutterApi {
   private final InstanceManager instanceManager;
 
-  public ZoomStateFlutterApiImpl(
-      BinaryMessenger binaryMessenger, InstanceManager instanceManager) {
+  public ZoomStateFlutterApiImpl(BinaryMessenger binaryMessenger, InstanceManager instanceManager) {
     super(binaryMessenger);
     this.instanceManager = instanceManager;
   }
 
-  /** Creates a {@link ZoomState} on the Dart side with its minimum zoom ratio and maximum zoom ratio. */
+  /**
+   * Creates a {@link ZoomState} on the Dart side with its minimum zoom ratio and maximum zoom
+   * ratio.
+   */
   void create(@NonNull ZoomState zoomState, @NonNull Reply<Void> reply) {
-    final float minZoomRatio = zoomState.getMinZoomRatio();
-    final float maxZoomRatio = zoomState.getMaxZoomRatio();
+    final Float minZoomRatio = zoomState.getMinZoomRatio();
+    final Float maxZoomRatio = zoomState.getMaxZoomRatio();
 
     if (!instanceManager.containsInstance(zoomState)) {
-    create(instanceManager.addHostCreatedInstance(zoomState),(Long) minZoomRatio, (Long) maxZoomRatio, reply);
+      create(
+          instanceManager.addHostCreatedInstance(zoomState),
+          minZoomRatio.doubleValue(),
+          maxZoomRatio.doubleValue(),
+          reply);
     }
   }
 }
